+++ frontend/app/src/entities/events/ui/node-details-events.tsx
